+++ base/forms/learning_unit/attribution_charge_repartition.py
@@ -49,14 +49,16 @@ class AttributionForm(forms.ModelForm):
         self.learning_unit_year = kwargs.pop("learning_unit_year")
         super().__init__(*args, **kwargs)
 
-        if not self.learning_unit_year.learning_container_year.container_type == learning_container_year_types.COURSE:
+        if self.learning_unit_year.learning_container_year.container_type != learning_container_year_types.COURSE or \
+            self.learning_unit_year.is_partim():
             del self.fields["start_year"]
             del self.fields["duration"]
             self.fields["function"].choices = Functions.choices_without_professor()
 
     def save(self, commit=True):
         instance = super().save(commit=False)
-        if not self.learning_unit_year.learning_container_year.container_type == learning_container_year_types.COURSE:
+        if self.learning_unit_year.learning_container_year.container_type == learning_container_year_types.COURSE and \
+            not self.learning_unit_year.is_partim():
             instance.end_year = instance.start_year + self.cleaned_data["duration"] - 1
 
         if commit:
